refactor(express-app): update express app to suport changes in feature icebreaker-messages-request
