+++ server/src/apps/express-server/controllers/index.ts
@@ -1,3 +1,4 @@
 export { fakeErrorHandler } from './fake-error.controller'
 export { fakeRandomHandler } from './fake-random.controller'
 export { fakeSuccessHandler } from './fake-success.controller'
+export { icebreakerMessagesHandler } from './icebreaker-messages.controller'
